+++ server/src/holidays/holidays.service.ts
@@ -11,31 +11,13 @@ import {
 import { PTO } from '../model/pto.entity';
 import { User } from '../model/user.entity';
 import { ErrorMessage, PTOInfo } from '../utils/types';
+import DateUtil from '../utils/Utils';
 
 @Injectable()
 export class HolidaysService {
   constructor(
     @InjectRepository(Holiday) private holidayRepo: Repository<Holiday>,
-    @InjectRepository(PTO) private PTORepo: Repository<PTO>,
-    @InjectRepository(User) private userRepo: Repository<User>,
   ) {}
-
-  getDatesBetweenDates = ({ startingDate, endingDate }: HolidayPeriod) => {
-    let dates = [];
-
-    const theStartDate = new Date(startingDate);
-    const theEndDate = new Date(endingDate);
-    while (theStartDate <= theEndDate) {
-      const year = new Date(theStartDate).getFullYear().toString();
-      let month = (new Date(theStartDate).getMonth() + 1).toString();
-      let day = new Date(theStartDate).getDate().toString();
-      month = month.length === 1 ? `0${month}` : month;
-      day = day.length === 1 ? `0${day}` : day;
-      dates = [...dates, `${year}-${month}-${day}`];
-      theStartDate.setDate(theStartDate.getDate() + 1);
-    }
-    return dates;
-  };
 
   getConstantHolidaysForTheCurrentYear = async (
     holidayPeriod: HolidayPeriod,
@@ -134,7 +116,7 @@ export class HolidaysService {
         startingDate: holidayPeriod.startingDate.toString(),
         endingDate: holidayPeriod.endingDate.toString(),
       };
-      const datesBetween = this.getDatesBetweenDates(holidayPeriodAsString);
+      const datesBetween = DateUtil.getDatesBetweenDates(holidayPeriodAsString);
       const constantHolidays = await this.getConstantHolidaysForTheCurrentYear(
         holidayPeriodAsString,
       );
@@ -157,157 +139,6 @@ export class HolidaysService {
       return datesWithAllHolidaysAndWeekends;
     } catch (error) {
       throw new Error(error.message);
-    }
-  }
-
-  saveHolidayIntoPTO = async (
-    holidayInfo: HolidayInfoDto,
-    user: User,
-  ): Promise<PTO> => {
-    try {
-      const approversProm = holidayInfo.approvers.map(async (el) => {
-        return await this.userRepo.findOne({ email: el });
-      });
-      const approvers = await Promise.all(approversProm);
-      const approoversValidation = approvers.indexOf(undefined);
-      if (approoversValidation >= 0) {
-        throw new Error(
-          `User with email ${holidayInfo.approvers[approoversValidation]} does not exist.`,
-        );
-      }
-
-      const employee = this.userRepo.create(user);
-      const newHoliday = this.PTORepo.create({
-        from_date: holidayInfo.startingDate.toString(),
-        to_date: holidayInfo.endingDate.toString(),
-        comment: holidayInfo.comment,
-        status: 'requested',
-        employee,
-        approvers,
-      });
-      return await this.PTORepo.save(newHoliday);
-    } catch (error) {
-      throw new Error(error.message);
-    }
-  };
-
-  validateHolidayPeriod = async (
-    holidayInfo: HolidayInfoDto,
-    user: User,
-  ): Promise<void> => {
-    try {
-      if (holidayInfo.startingDate > holidayInfo.endingDate) {
-        throw new Error('The first date must not be after the last date!');
-      }
-
-      const vacationDays = await this.calculateDays({
-        startingDate: holidayInfo.startingDate,
-        endingDate: holidayInfo.endingDate,
-      });
-
-      let isThereAWorkdayInSubmittedPeriod = false;
-
-      for (let i = 0; i < vacationDays.length; i++) {
-        if (vacationDays[i].status === 'workday') {
-          isThereAWorkdayInSubmittedPeriod = true;
-          break;
-        }
-      }
-      if (!isThereAWorkdayInSubmittedPeriod) {
-        throw new Error('There are not working days in the submitted period.');
-      }
-
-      const employeeHolidays = await this.PTORepo.find({
-        where: [
-          { employee: user.id, status: 'requested' },
-          { employee: user.id, status: 'approved' },
-        ],
-      });
-
-      let overlapIndex = -1;
-
-      for (let i = 0; i < employeeHolidays.length; i++) {
-        if (
-          !(
-            holidayInfo.endingDate < employeeHolidays[i].from_date ||
-            holidayInfo.startingDate > employeeHolidays[i].to_date
-          )
-        ) {
-          overlapIndex = i;
-          break;
-        }
-      }
-      if (overlapIndex >= 0) {
-        throw new Error(
-          `The period you submitted is overlaping with another vacation from ${employeeHolidays[overlapIndex].from_date} to ${employeeHolidays[overlapIndex].to_date}`,
-        );
-      }
-    } catch (error) {
-      throw new Error(error.message);
-    }
-  };
-
-  public async postHoliday(
-    holidayInfo: HolidayInfoDto,
-    user: User,
-  ): Promise<PTO | ErrorMessage> {
-    try {
-      await this.validateHolidayPeriod(holidayInfo, user);
-      return await this.saveHolidayIntoPTO(holidayInfo, user);
-    } catch (error) {
-      throw new Error(error.message);
-    }
-  }
-
-  public async getUserPTOs(user: User): Promise<Array<PTOInfo>> {
-    try {
-      const userHolidays = await this.PTORepo.find({
-        where: { employee: user.id },
-      });
-
-      const userHolidaysWithCalculatedPTOs = userHolidays.map(async (el) => {
-        const eachDayStatus = await this.calculateDays({
-          startingDate: el.from_date,
-          endingDate: el.to_date,
-        });
-        if (Array.isArray(eachDayStatus)) {
-          const workDays = eachDayStatus.filter(
-            (elem) => elem.status === 'workday',
-          );
-          return {
-            ...el,
-            totalDays: eachDayStatus.length,
-            PTODays: workDays.length,
-          };
-        }
-      });
-      const resolvedHolidaysInfo = await Promise.all(
-        userHolidaysWithCalculatedPTOs,
-      );
-      return resolvedHolidaysInfo;
-    } catch (error) {
-      throw new Error('Something went wrong with getting user holidays');
-    }
-  }
-
-  private async getPTOFullInfo(PTOId: string): Promise<PTO> {
-    return await this.PTORepo.findOne({
-      where: { id: PTOId },
-      relations: ['employee', 'approvers'],
-    });
-  }
-
-  public async getPTOById(PTOId: string): Promise<PTOFullInfo> {
-    try {
-      const PTOInfo = await this.getPTOFullInfo(PTOId);
-      const eachDayStatus = await this.calculateDays({
-        startingDate: PTOInfo.from_date,
-        endingDate: PTOInfo.to_date,
-      });
-      const PTOInfoWithEachDayStatus = { ...PTOInfo, eachDayStatus };
-      return PTOInfoWithEachDayStatus;
-    } catch (error) {
-      throw new Error('Invalid PTO.');
     }
   }
 }
